Update: Made a fix to the cluster assignments
Fixed logic.  Within the logic, where topics were assigned to cluster or
groups, initially when the cluster were new, it would add the cluster
but skip adding the topic. Hence, the initial topics would not be
assigned clusters. Fixed, this such that when adding new clusters it
also add the topics.